Use "query parameter" only on the CRI side.
Avoid undefined use of "query parameter" on the URI side, where we
instead now make explicit how the conversion into CRIs leads to deriving query
parameters from the URI-side query strings.

(MB: "...the core issue is that "query parameters" is not actually a
defined term for URIs, even though we all "know" what they are by
convention. So if we're clear that we're defining the term "query
parameters" for CRIs to mean the elements of this array, then the term
is fine throughout the document. But then we need to not use that term
to refer to the things between ampersands in URI query strings.")